backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pandas
python-dotenv
google-generativeai
openai
sqlparse
pymysql
psycopg2-binary
requests
httpx
